+++ includes/draw.h
@@ -6,15 +6,17 @@
 /*   By: tsishika <tsishika@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/12 16:11:24 by tsishika          #+#    #+#             */
-/*   Updated: 2024/03/15 22:41:30 by tsishika         ###   ########.fr       */
+/*   Updated: 2024/03/15 22:51:34 by tsishika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef DRAW_H
 # define DRAW_H
 
+# include "hook.h"
 # include "struct.h"
 # include "utils.h"
+
 # include "mlx.h"
 
 // draw.c
