+++ includes/hook.h
@@ -1,31 +1,25 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   checker.h                                          :+:      :+:    :+:   */
+/*   hook.h                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: tsishika <tsishika@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/03/11 21:33:07 by tsishika          #+#    #+#             */
-/*   Updated: 2024/03/15 22:50:20 by tsishika         ###   ########.fr       */
+/*   Created: 2024/03/15 22:46:43 by tsishika          #+#    #+#             */
+/*   Updated: 2024/03/15 22:51:23 by tsishika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef CHECKER_H
-# define CHECKER_H
+#ifndef HOOK_H
+# define HOOK_H
 
-# include "stdbool.h"
+# include "stdlib.h"
 
 # include "struct.h"
-# include "utils.h"
-# include "parser.h"
+# include "mlx.h"
 
-// check_closed.c
-void	validate_map_walkability(t_data *data);
+int	closed_window(t_mlx *mlx);
 
-// check_map.c
-void	check_input_data(t_data *data);
-
-// check_player.c
-void	scan_map_and_check_player(t_data *data);
+int	key_press(int keycode, t_data *data);
 
 #endif
