+++ srcs/draw/draw.c
@@ -6,29 +6,11 @@
 /*   By: tsishika <tsishika@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/12 16:11:03 by tsishika          #+#    #+#             */
-/*   Updated: 2024/03/15 15:33:25 by tsishika         ###   ########.fr       */
+/*   Updated: 2024/03/15 22:48:03 by tsishika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "draw.h"
-
-int	closed_window(t_mlx *mlx)
-{
-	mlx_clear_window(mlx->mlx, mlx->win);
-	mlx_destroy_window(mlx->mlx, mlx->win);
-	exit(0);
-}
-
-int	key_press(int keycode, t_data *data)
-{
-	if (keycode == 53)
-	{
-		mlx_clear_window(data->mlx->mlx, data->mlx->win);
-		mlx_destroy_window(data->mlx->mlx, data->mlx->win);
-		exit(0);
-	}
-	return (0);
-}
 
 	// その他のmlx系の初期化
 	// mlx_loop_hookとかで描画
