+++ srcs/draw/hook/closed_window.c
@@ -1,31 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   checker.h                                          :+:      :+:    :+:   */
+/*   closed_window.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: tsishika <tsishika@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/03/11 21:33:07 by tsishika          #+#    #+#             */
-/*   Updated: 2024/03/15 22:50:20 by tsishika         ###   ########.fr       */
+/*   Created: 2024/03/15 22:45:32 by tsishika          #+#    #+#             */
+/*   Updated: 2024/03/15 22:51:01 by tsishika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef CHECKER_H
-# define CHECKER_H
+#include "hook.h"
 
-# include "stdbool.h"
-
-# include "struct.h"
-# include "utils.h"
-# include "parser.h"
-
-// check_closed.c
-void	validate_map_walkability(t_data *data);
-
-// check_map.c
-void	check_input_data(t_data *data);
-
-// check_player.c
-void	scan_map_and_check_player(t_data *data);
-
-#endif
+int	closed_window(t_mlx *mlx)
+{
+	mlx_clear_window(mlx->mlx, mlx->win);
+	mlx_destroy_window(mlx->mlx, mlx->win);
+	exit(0);
+}
