+++ srcs/draw/hook/key_press.c
@@ -1,31 +1,24 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   checker.h                                          :+:      :+:    :+:   */
+/*   key_press.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: tsishika <tsishika@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/03/11 21:33:07 by tsishika          #+#    #+#             */
-/*   Updated: 2024/03/15 22:50:20 by tsishika         ###   ########.fr       */
+/*   Created: 2024/03/15 22:45:57 by tsishika          #+#    #+#             */
+/*   Updated: 2024/03/15 22:51:03 by tsishika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef CHECKER_H
-# define CHECKER_H
+#include "hook.h"
 
-# include "stdbool.h"
-
-# include "struct.h"
-# include "utils.h"
-# include "parser.h"
-
-// check_closed.c
-void	validate_map_walkability(t_data *data);
-
-// check_map.c
-void	check_input_data(t_data *data);
-
-// check_player.c
-void	scan_map_and_check_player(t_data *data);
-
-#endif
+int	key_press(int keycode, t_data *data)
+{
+	if (keycode == 53)
+	{
+		mlx_clear_window(data->mlx->mlx, data->mlx->win);
+		mlx_destroy_window(data->mlx->mlx, data->mlx->win);
+		exit(0);
+	}
+	return (0);
+}
