+++ srcs/main.c
@@ -6,7 +6,7 @@
 /*   By: tsishika <tsishika@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/27 16:48:40 by tsishika          #+#    #+#             */
-/*   Updated: 2024/03/15 22:33:12 by tsishika         ###   ########.fr       */
+/*   Updated: 2024/03/15 22:51:59 by tsishika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,9 +55,7 @@ int	main(int argc, char **argv)
 }
 
 #ifdef DEBUG
-
 __attribute__((destructor)) static void destructor() {
 	system("leaks -q cub3D");
 }
-
 #endif
